1.Ajax显示三级分类

+++ src/cn/happy/servlet/SetCategoriesServlet.java
@@ -1,7 +1,7 @@
 package cn.happy.servlet;
 
+import cn.happy.bean.Easybuy_product_parent;
 import cn.happy.service.ICategoryService;
-import cn.happy.service.ISliderService;
 import cn.happy.service.impl.CategoryServiceImpl;
 import org.apache.log4j.Logger;
 
@@ -12,6 +12,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.File;
 import java.io.IOException;
+import java.util.List;
 
 /**
  * Created by master on 17-8-27.
@@ -28,26 +29,42 @@ public class SetCategoriesServlet extends HttpServlet {
             doShow(request, response);
             return;
         }
+        if (action != null && action.equals("delete")) {
+            doDel(request, response);
+            return;
+        }
+    }
+
+    private void doDel(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        ICategoryService service = new CategoryServiceImpl();
+        String epp_id = request.getParameter("epp_id");
+        deleteExpiredFile(service, epp_id);
+        if (service.deleteCategoriesByid(epp_id)) {
+            response.sendRedirect("/easybuy/AdminServlet/SetCategoriesServlet?action=show");
+        }
+    }
+
+    private void doShow(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        ICategoryService service = new CategoryServiceImpl();
+        List<Easybuy_product_parent> parents = service.getParents();
+        request.setAttribute("parents", parents);
+        request.getRequestDispatcher("/background_jsp/category.jsp").forward(request, response);
+    }
+
+    private void deleteExpiredFile(ICategoryService service, String id) {
+        String fileName = service.getImageByParentId(id);
+        String leftPath = getServletContext().getRealPath("/images/");
+        if (fileName != null && !fileName.equals("")) {
+            File file = new File(leftPath, fileName);
+            if (file.exists() && file.isFile()) {
+                if (!file.delete()) {
+                    logger.error("Parent img delete failed-->" + leftPath + "/" + fileName);
+                }
+            }
+        }
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         doPost(request, response);
-    }
-
-    private void doShow(HttpServletRequest request, HttpServletResponse response) {
-        ICategoryService service = new CategoryServiceImpl();
-
-
-    }
-
-    private void deleteExpiredFile(ISliderService service, String id) {
-        String fileName = service.getImageBySliderId(id);
-        String leftPath = getServletContext().getRealPath("/images/");
-        File file = new File(leftPath, fileName);
-        if (file.exists() && file.isFile()) {
-            if (!file.delete()) {
-                logger.error("file delete failed-->" + leftPath + "/" + fileName);
-            }
-        }
     }
 }
